Refactor source layout
This introduces a module core that takes care of deciding how to handle
messages, and moves the websocket and web stuff into modules by that
name.

Instead of passing the "app" value to these modules (as was done with
database), the imports of these secondary modules is moved down so they
can import the app value (without the circular reference breaking
things).