Panic-mode implementation
Added enable/disable methods for levelController and flowController classes.
When a panic signal from any controller is passed to the battery doCycle method all pumps are disabled.
On receiving a new configuration from the web app the battery is re-enabled.

Direct write access to the pumpPins is restricted to the setPumpState and disable methods.
All other code should use setPumpState.