Add authentication support
This adds the Flask-User framework, which enables login, invites,
registration, password-changing.

The base template is modified to show the login status and a login link,
and it adds a bit of styling at the same time (this is the base template
copied from Flask-User, except that the "User Profile" link was removed
(since it is also reachable by clicking the username).

The login status is not yet applied to the interface drawn, and
authentication is not yet required for the websocket connection.

The users are stored in the same database, except using SqlAlchemy
instead of a direct SQL connection, so it might be good to convert the
other database accesses to use SqlAlchemy as well later.