Reorder PID initialization
The PID library uses the initial output value to initialize the integral
sum. Previously, this pointed to some uninitialized memory, leading to
unexpected behaviour. Now, the output value is cleared before
initializing the PID, so the integral sum starts at 0.

Additionally, `setMode(AUTOMATIC)` is called last, which simplifies
other initialization in the library.