FlowController: Support positive and negative flow
This uses a direction relay output on the flow controller to attribute
pulses to either positive or negative flow, keeping a separate counter
for each.

Currently, this reverse flow is not yet used anywhere yet, though it
might make the reported flow value more accurate (if the current meter
was configured to also generate pulses for reverse flow, which would be
counted as forward flow until now).