FlowController: Correct for flow after disabling pump
When the target flow is reached, the pump is shutdown. Any flow occuring
after that was counted and reported, but did not influence the pump
control at all. With this commit, any extra flow (or reverse flow)
occuring after the pump is shutdown will cause a correction on the
target count for the next cycle, so the average flow should be better.