FlowController: Fix rounding error in the reported flow
In the flow calculation, the actual cycle time would be used rather than
the configured cycle time. However, a part of that calculation would be
intermediately rounded, causing a value of e.g. 11.99 cycles per hour to
be rounded to 11 cycles per hour.

Effectively, this means that reported flow values would be around 9%
(1/12th) lower that their real values.

This is fixed by doing this calculation using floating point numbers.