FlowController: Prevent accumulating backlog
Previously, when the realized flow would lag behind the target flow
for multiple cycles, this backlog could accumulate and eventually
overflow. This should not normally happen, but it could happen when the
flow is set too high, or something obstructs the flow (or the pump is
broken). Now, any missed flow is not included in the correction (but any
*reverse* flow *is* taken into account still).